requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1
pyyaml>=5.4.1
python-dotenv>=0.19.0

# Data processing
pyarrow>=6.0.1
dask>=2021.9.1
openpyxl>=3.0.9
fastparquet>=0.7.2
xlrd>=2.0.1

# Forecasting
nixtlats>=0.1.0
prophet>=1.1.1
statsmodels>=0.13.0
pmdarima>=1.8.5

# News collection
praw>=7.4.0
pytrends>=4.7.3
tradingeconomics>=0.2.0
textblob>=0.15.3
beautifulsoup4>=4.10.0
requests>=2.26.0

# Visualization
nbformat>=5.1.3
jupyter>=1.0.0
ipykernel>=6.4.1
notebook>=6.4.4

# Web presentation
streamlit

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.1

# Development
black>=21.9b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.3

# Langchain ecosystem
langchain

# Langgraph and Autogen
langgraph
autogen

# Vector stores and retrieval
chromadb
faiss-cpu
cassio

# Presentation generation
python-pptx>=0.11.0

# API framework
fastapi
uvicorn
sse_starlette

# Transformers & embeddings
sentence_transformers

# Azure Identity
azure-identity==1.15.0
azure-core==1.30.0